Code refactoring 🚀🔨
