+++ src/ignitus-HomePage/ignitus-GetStarted/Components/GetStarted.test.js
