+++ src/ignitus-HomePage/ignitus-GetStarted/actionTypes.js
