+++ src/ignitus-HomePage/ignitus-GetStarted/reducers.js
